Move methods to optics package and add Reverse suffix
This moves the deriving methods to the `optics` package and put a `Reverse` suffix on the methods that currently have a dummy implicit.
