+++ modules/optics/src/test/scala/io/circe/optics/DeriveWithIsoSuite.scala
@@ -2,16 +2,15 @@ package io.circe.optics
 
 import io.circe._
 import io.circe.generic.semiauto._
-import io.circe.optics.semiauto._
 import io.circe.syntax._
 import io.circe.tests.CirceSuite
 import monocle.Iso
 
-class SemiautoSuite extends CirceSuite {
+class DeriveWithIsoSuite extends CirceSuite {
 
   case class UserSnake(id: Long, first_name: String, last_name: String)
   case class UserCamel(id: Long, firstName: String, lastName: String)
-  
+
   val snake2camel = Iso[UserSnake, UserCamel] {
     (s: UserSnake) => UserCamel(s.id, s.first_name, s.last_name)
   } {
@@ -42,20 +41,20 @@ class SemiautoSuite extends CirceSuite {
     assert(camelEncoder(UserCamel(1, "John", "Doe")) == john)
   }
 
-  "deriveDecoderWithIso[B, A]" should "create a decoder[B] with an Iso[B, A]" in {
+  "deriveDecoderWithIsoReverse[B, A]" should "create a decoder[B] with an Iso[B, A]" in {
     implicit val iso: Iso[UserCamel, UserSnake] = snake2camel.reverse
     implicit val snakeDecoder: Decoder[UserSnake] = deriveDecoder[UserSnake]
 
-    val camelDecoder: Decoder[UserCamel] = deriveDecoderWithIso[UserCamel, UserSnake]
+    val camelDecoder: Decoder[UserCamel] = deriveDecoderWithIsoReverse[UserCamel, UserSnake]
 
     assert(camelDecoder.decodeJson(john) == Right(UserCamel(1, "John", "Doe")))
   }
 
-  "deriveEncoderWithIso[B, A]" should "create an encoder[B] with an Iso[B, A]" in {
+  "deriveEncoderWithIsoReverse[B, A]" should "create an encoder[B] with an Iso[B, A]" in {
     implicit val iso: Iso[UserCamel, UserSnake] = snake2camel.reverse
     implicit val snakeEncoder: Encoder[UserSnake] = deriveEncoder[UserSnake]
 
-    val camelEncoder: Encoder[UserCamel] = deriveEncoderWithIso[UserCamel, UserSnake]
+    val camelEncoder: Encoder[UserCamel] = deriveEncoderWithIsoReverse[UserCamel, UserSnake]
 
     assert(camelEncoder(UserCamel(1, "John", "Doe")) == john)
   }
